+++ src/main/java/com/nexters/phochak/auth/presentation/AuthController.java
@@ -1,8 +1,8 @@
 package com.nexters.phochak.auth.presentation;
 
+import com.nexters.phochak.auth.application.AuthService;
 import com.nexters.phochak.auth.application.JwtTokenService;
 import com.nexters.phochak.post.CommonResponse;
-import com.nexters.phochak.user.application.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,12 +16,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/v2/auth")
 @RestController
 public class AuthController {
-    private final UserService userService;
+    private final AuthService authService;
     private final JwtTokenService jwtTokenService;
 
     @GetMapping("/login/{provider}")
     public CommonResponse<JwtResponseDto> loginV2(@PathVariable String provider, @Valid LoginRequestDto requestDto) {
-        Long loginUserId = userService.login(provider, requestDto);
+        Long loginUserId = authService.login(provider, requestDto);
         return new CommonResponse<>(jwtTokenService.issueToken(loginUserId));
     }
 }
