+++ src/main/java/com/nexters/phochak/user/application/UserService.java
@@ -1,6 +1,5 @@
 package com.nexters.phochak.user.application;
 
-import com.nexters.phochak.auth.presentation.LoginRequestDto;
 import com.nexters.phochak.ignore.IgnoredUserResponseDto;
 import com.nexters.phochak.user.UserCheckResponseDto;
 import com.nexters.phochak.user.UserInfoResponseDto;
@@ -8,14 +7,6 @@ import com.nexters.phochak.user.UserInfoResponseDto;
 import java.util.List;
 
 public interface UserService {
-
-    /**
-     * OAuth 로그인을 진행한다.
-     *
-     * @param provider
-     * @param requestDto
-     */
-    Long login(String provider, LoginRequestDto requestDto);
 
     /**
      * 해당 id의 유저 정보를 조회한다.
